MAJ SHOW_TAG 2

+++ src/main/resources/db/migration/V18_1__create_show_tag_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE show_tag
 (
     show_id INT(11) NOT NULL,
-    tag_id  INT(11) NOT NULL,
+    tag_id  BIGINT  NOT NULL,
     PRIMARY KEY (show_id, tag_id),
     CONSTRAINT fk_show FOREIGN KEY (show_id)
         REFERENCES shows (id)
